requirements.txt
Django==3.0
djangorestframework==3.10.3
django-environ==0.4.5
psycopg2==2.8.4
trafaret==2.0.1